+++ packages/@runlightyear/github/src/webhooks/payloads/asPushPayload.ts
@@ -1,5 +1,6 @@
 import { WebhookDeliveryData } from "@runlightyear/lightyear";
 import commonPayload, { CommonPayload } from "./commonPayload";
+import { PushCommit } from "../../types/PushCommit";
 
 /**
  * Documentation: https://docs.github.com/webhooks-and-events/webhooks/webhook-events-and-payloads#push
@@ -32,7 +33,7 @@ export interface PushPayload extends CommonPayload {
   /**
    * 	For pushes where after is or points to a commit object, an expanded representation of that commit. For pushes where after refers to an annotated tag object, an expanded representation of the commit pointed to by the annotated tag.
    */
-  head_commit: object;
+  headCommit: PushCommit;
   /**
    * 	URL that shows the changes in this ref update, from the before commit to the after commit. For a newly created ref that is directly based on the default branch, this is the comparison between the head of the default branch and the after commit. Otherwise, this shows all commits until the after commit.
    */
@@ -40,53 +41,7 @@ export interface PushPayload extends CommonPayload {
   /**
    * 	An array of commit objects describing the pushed commits. (Pushed commits are all commits that are included in the compare between the before commit and the after commit.)
    */
-  commits: Array<{
-    /**
-     * The SHA of the commit.
-     */
-    id: string;
-    /**
-     * The ISO 8601 timestamp of the commit.
-     */
-    timestamp: string;
-    /**
-     * The commit message.
-     */
-    message: string;
-    /**
-     * The git author of the commit.
-     */
-    author: {
-      /**
-       * The git author's name.
-       */
-      name: string;
-      /**
-       * The git author's email address.
-       */
-      email: string;
-    };
-    /**
-     * URL that points to the commit API resource.
-     */
-    url: string;
-    /**
-     * Whether this commit is distinct from any that have been pushed before.
-     */
-    distinct: boolean;
-    /**
-     * An array of files added in the commit. For extremely large commits where GitHub is unable to calculate this list in a timely manner, this may be empty even if files were added.
-     */
-    added: string[];
-    /**
-     * An array of files modified by the commit. For extremely large commits where GitHub is unable to calculate this list in a timely manner, this may be empty even if files were modified.
-     */
-    modified: string[];
-    /**
-     * An array of files removed in the commit. For extremely large commits where GitHub is unable to calculate this list in a timely manner, this may be empty even if files were removed.
-     */
-    removed: string[];
-  }>;
+  commits: Array<PushCommit>;
   /**
    * The user who pushed the commits.
    */